ALL initial mechanics to replace the sim::ParticleList with an art/LAr class that is art::Ptr aware.
The ParticleList is still provided in this version, as it has
hooks everywhere which I will need to remove and update for the new
struct before I can completely eliminate the particle list. This does
mean that for now the BackTracking Suite is VERY memory innefficient as
it maintains the complete particle record in two separate instances.