Larsoft Issue #22339 I am revisiting this issue after a report from ProtoDUNE that they see this message often with high energy tracks.
With the change in OpFastScintillation to write each photon to the
OpDetBacktrackerRecords, the energy was being incorrectly duplicated in
the BTRs in ticks with multiple photons from the same G4Step.
This was corrected in my commit last week by setting the Edeposit to 0
after the first photon was inserted, but, this caused subsequent
photons to be written with 0 energy, causing the error message that
appeared so prolifically.

Now OpFastScintillation will instead take the total energy provided by
the ISCalculator for the given step, and will divide that energy into
equal parts for each photon to be written. This results in the same
value being written, as well as making sure no zero values should ever
be passed to the BTRs.